+++ src/test/run.ts
@@ -12,6 +12,7 @@ import "./getRandomProblemByFilter.test";
 import "./statusSelect.test";
 import "./randomProblemDisplay.test";
 import "./tagsInput.test";
+import "./getRatingDistribution.test";
 import { run } from "./testKit";
 
 await run();